fix: move call detail root link into its own file
