+++ clients/web/src/lib/calls/get-calls-tab-base-path.ts
@@ -1,0 +1,7 @@
+import { useLocation } from "@solidjs/router";
+import { getRootPathByUrlSegment } from "../formatters";
+
+export function getCallsTabBasePath() {
+  const { pathname } = useLocation();
+  return getRootPathByUrlSegment(pathname, "calls");
+}
