+++ clients/web/src/lib/formatters.ts
@@ -46,3 +46,14 @@ export function getDetailedTime(date: Date) {
     fractionalSecondDigits: 3,
   }).format(date);
 }
+
+export function getRootPathByUrlSegment(path: string, segment: string) {
+  return path
+    .split("/")
+    .slice(
+      0,
+      path.split("/").findIndex((e) => e === segment),
+    )
+    .concat(segment)
+    .join("/");
+}
