+++ clients/web/src/lib/tauri/get-tauri-tab-base-path.ts
@@ -1,17 +1,7 @@
 import { useLocation } from "@solidjs/router";
+import { getRootPathByUrlSegment } from "../get-root-path-by-url-segment";
 
 export function getTauriTabBasePath() {
   const { pathname } = useLocation();
   return getRootPathByUrlSegment(pathname, "tauri");
-}
-
-export function getRootPathByUrlSegment(path: string, segment: string) {
-  return path
-    .split("/")
-    .slice(
-      0,
-      path.split("/").findIndex((e) => e === segment),
-    )
-    .concat(segment)
-    .join("/");
 }
